Feat: split cypress commands according to the feature (#279)
* feat: split cypress commands according to the feature

* refactor: automated test for adding delta meter reads

* refactor: automated tests for account settings and add aggregate meter reads

* refactor: automated test for add devices

* refactor: test cases for adding history meter reads

* refactor: test cases for adding reservations

* refactor: test cases for admin login

* refactor: test cases for buyer user login

* refactor: test cases for buyer user permission setup

* refactor: test cases for buyer user sign up

* refactor: automated test cases

* refactor: automated test cases

* refactor: automated test cases

* refactor: test cases for certificates

* refactor: automated test cases

* refactor: automated test cases

---------
