+++ cypress/e2e/04-developer-user-signup.cy.ts
@@ -1,7 +1,6 @@
 describe('Sign up Test', () => {
   beforeEach(() => {
     cy.wait(5000);
-    // cy.clearDatabase();
   });
 
   it('should sign up a new developer user', () => {
