+++ cypress/e2e/05-buyer-user-signup.cy.ts
@@ -1,7 +1,6 @@
 describe('Sign Up Test', () => {
   beforeEach(() => {
     cy.wait(5000);
-    // cy.clearDatabase();
   });
   it('should sign up a new buyer user', () => {
     cy.buyerUserSignup();
