+++ cypress/e2e/08-add-delta-meter-read.cy.ts
@@ -1,10 +1,9 @@
 describe('Add delta meter Read', () => {
-    beforeEach(() => {
-      cy.wait(1000);
-    });
-    it('should add delta meter Read', () => {
-      cy.developerUserLogin().wait(20000);
-      cy.addDeltaMeterRead();
-    });
+  beforeEach(() => {
+    cy.wait(1000);
   });
-  
+  it('should add delta meter Read', () => {
+    cy.developerUserLogin().wait(20000);
+    cy.addDeltaMeterRead();
+  });
+});
