+++ cypress/e2e/09-add-aggregate-meter-read.cy.ts
@@ -1,10 +1,9 @@
 describe('Add aggregate Meter Read', () => {
-    beforeEach(() => {
-      cy.wait(1000);
-    });
-    it('should add aggregate meter Read', () => {
-      cy.developerUserLogin().wait(20000);
-      cy.addAggregateMeterRead();
-    });
+  beforeEach(() => {
+    cy.wait(1000);
   });
-  
+  it('should add aggregate meter Read', () => {
+    cy.developerUserLogin().wait(20000);
+    cy.addAggregateMeterRead();
+  });
+});
