+++ cypress/e2e/12-certificate-filtering.cy.ts
@@ -1,10 +1,9 @@
 describe('Filter certificates', () => {
-    beforeEach(() => {
-      cy.wait(1000);
-    });
-    it('should filter certificates', () => {
-      cy.buyerUserLogin().wait(20000);
-      cy.certificateFilter();
-    });
+  beforeEach(() => {
+    cy.wait(1000);
   });
-  
+  it('should filter certificates', () => {
+    cy.buyerUserLogin().wait(20000);
+    cy.certificateFilter();
+  });
+});
