+++ cypress/support/commands/developer-user-signup.ts
@@ -1,0 +1,27 @@
+import 'cypress-file-upload';
+const UI_BASE_URL = Cypress.env('UI_BASE_URL');
+
+Cypress.Commands.add('developerUserSignup', function () {
+  cy.fixture('developer-user-signup.json').then((data) => {
+    cy.visit(`${UI_BASE_URL}/login`).wait(1000);
+    cy.get('[test-id="register"]').click();
+    data.forEach((step) => {
+      switch (step.action) {
+        case 'type':
+          return cy
+            .get(step.selector)
+            .type(step.value)
+            .should('have.value', step.value);
+        case 'click':
+          return cy.get(step.selector).click().wait(1000);
+        case 'select':
+          return cy
+            .get(step.selector)
+            .click()
+            .then(() => {
+              cy.get('mat-option').contains(step.value).click();
+            });
+      }
+    });
+  });
+});
