+++ cypress/support/commands/meter-read-bulk-upload.ts
@@ -1,0 +1,16 @@
+Cypress.Commands.add('meterReadBulkUpload', function () {
+  cy.fixture('meter-read-bulk-upload.json').then((data) => {
+    data.forEach((step) => {
+      switch (step.action) {
+        case 'click':
+          return cy.get(step.selector).click().wait(1000);
+        case 'upload':
+          return cy
+            .get(step.selector)
+            .attachFile('files/d-rec_bulk_upload_meter_read_template.csv', {
+              force: true,
+            });
+      }
+    });
+  });
+});
